// src/payload-types.ts
export interface Lesson {
  id: number;
  title: string;
  /**
   * Which education level this lesson is for.
   */
  level: 'elementary' | 'middle' | 'high';
  /**
   * e.g. Mathematics, History, Biology
   */
  subject?: string | null;
  description?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}